'warp elasticsearch --clear-all' turns into 'warp elasticsearch flush'. Help function added to cmd. Fix in 'warp elasticsearch info'

+++ .warp/bin/elasticsearch.sh
@@ -15,7 +15,9 @@ function elasticsearch_info()
     ES_HOST="elasticsearch"
     ES_VERSION=$(warp_env_read_var ES_VERSION)
     ES_MEMORY=$(warp_env_read_var ES_MEMORY)
-    ES_HOST2CONTAINER_PORT=$(docker inspect --format='{{(index (index .NetworkSettings.Ports "9200/tcp") 0).HostPort}}' $(warp docker ps -q elasticsearch))
+    if [ $(warp_check_is_running) = true ]; then
+        ES_HOST2CONTAINER_PORT=$(docker inspect --format='{{(index (index .NetworkSettings.Ports "9200/tcp") 0).HostPort}}' $(warp docker ps -q elasticsearch))
+    fi
 
     MODE_SANDBOX=$(warp_env_read_var MODE_SANDBOX)
 
@@ -85,33 +87,45 @@ elasticsearch-simil_ssh() {
             exit 0
         else
             elasticsearch-ssh_wrong_input
-            exit 1
         fi
     fi
 }
 
-elasticsearch-clear_all() {
+elasticsearch-flush() {
     : '
     This function unlocks and delete all indexes.
     '
 
-    # Parsing ES dynamic binded port:
-    ES_HOST2CONTAINER_PORT=$(docker inspect --format='{{(index (index .NetworkSettings.Ports "9200/tcp") 0).HostPort}}' $(warp docker ps -q elasticsearch))
-    # Unlocking indexes:
-    ACK=$(curl --no-progress-meter -X PUT -H "Content-Type: application/json" http://localhost:$ES_HOST2CONTAINER_PORT/_all/_settings -d '{"index.blocks.read_only_allow_delete": null}')
-    if [[ $(echo "$ACK" | grep '{"acknowledged":true}') ]]; then
-        warp_message "* Unlocking indexes... $(warp_message_ok [ok])"
-        # Deleting indexes:
-        ACK=$(curl --no-progress-meter -X DELETE "localhost:$ES_HOST2CONTAINER_PORT/_all")
+    if [[ $* == "-h" || $* == "--help" ]]; then
+        elasticsearch-flush_help
+        exit 0
+    elif [[ -z $* ]]; then
+        # Check if warp is running:    
+        if [ $(warp_check_is_running) = false ]; then
+            warp_message_error "The containers are not running"
+            warp_message_error "please, first run warp start"
+            exit 1
+        fi
+        # Parsing ES dynamic binded port:
+        ES_HOST2CONTAINER_PORT=$(docker inspect --format='{{(index (index .NetworkSettings.Ports "9200/tcp") 0).HostPort}}' $(warp docker ps -q elasticsearch))
+        # Unlocking indexes:
+        ACK=$(curl --no-progress-meter -X PUT -H "Content-Type: application/json" http://localhost:$ES_HOST2CONTAINER_PORT/_all/_settings -d '{"index.blocks.read_only_allow_delete": null}')
         if [[ $(echo "$ACK" | grep '{"acknowledged":true}') ]]; then
-            warp_message "* Deleting indexes... $(warp_message_ok [ok])"
+            warp_message "* Unlocking indexes... $(warp_message_ok [ok])"
+            # Deleting indexes:
+            ACK=$(curl --no-progress-meter -X DELETE "localhost:$ES_HOST2CONTAINER_PORT/_all")
+            if [[ $(echo "$ACK" | grep '{"acknowledged":true}') ]]; then
+                warp_message "* Deleting indexes... $(warp_message_ok [ok])"
+            else
+                warp_message_error "Delete process fail"
+                exit 1
+            fi
         else
-            warp_message_error "Delete process fail"
+            warp_message_error "Unlock process fail"
             exit 1
         fi
     else
-      warp_message_error "Unlock process fail"
-      exit 1
+        elasticsearch-flush_wrong_input
     fi
 }
 
@@ -123,21 +137,9 @@ function elasticsearch_main()
             elasticsearch_command $*  
         ;;
 
-        --clear-all)
+        flush)
             shift
-            # Check for wrong input:
-            if [[ -n $1 ]]; then
-                warp_message_error "Wrong input."
-                warp_message "Just use $(warp_message_warn "warp elasticsearch --clear-all")"
-                exit 1
-            fi
-            # Check if warp is running:    
-            if [ $(warp_check_is_running) = false ]; then
-                warp_message_error "The containers are not running"
-                warp_message_error "please, first run warp start"
-                exit 1
-            fi
-            elasticsearch_clear_all
+            elasticsearch-flush $*
         ;;
 
         info)
@@ -163,5 +165,11 @@ function elasticsearch_main()
 elasticsearch-ssh_wrong_input() {
     warp_message_error "Wrong input."
     elasticsearch-ssh_help
+    exit 1
+}
+
+elasticsearch-flush_wrong_input() {
+    warp_message_error "Wrong input."
+    elasticsearch-flush_help
     exit 1
 }
